api/pyproject.toml
[project]
name = "solomon-api"
version = "1.0.0"
description = "Solomon LLM Service API by 4grit"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.116.1",
    "uvicorn>=0.23.1",
    "gunicorn>=23.0.0",
    "httptools>=0.6.4",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.9",

    # HTTP & Serialization
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "orjson>=3.11.1",

    # Auth & Environment
    "python-jose>=3.4.0",
    "cryptography>=44.0.2",
    "python-dotenv>=1.1.0",

    # Database
    "motor>=3.7.0",
    "pinecone[grpc,asyncio]>=7.0.1",
    "pymilvus>=2.5.8",
    "hiredis>=3.1.0",
    "redis>=5.2.1",

    # Cloud
    "aioboto3>=15.5.0",

    # LLM Utilities
    "tiktoken>=0.9.0",
    "transformers>=4.51.1",
    "numpy>=2.2.4",

    # Converting
    "html2text>=2025.4.15",

    # MCP
    "fastmcp>=2.10.6",

    # Timezone
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
